+++ crates/installer/src/lib.rs
@@ -2,6 +2,7 @@ mod error;
 
 use futures::StreamExt;
 use starbase_archive::Archiver;
+use starbase_styles::color;
 use starbase_utils::fs::{self, FsError};
 use std::cmp;
 use std::env;
@@ -11,7 +12,7 @@ use std::io::Write;
 use std::path::{Path, PathBuf};
 use std::time::SystemTime;
 use system_env::SystemLibc;
-use tracing::instrument;
+use tracing::{instrument, trace};
 
 pub use error::ProtoInstallerError;
 
@@ -58,6 +59,13 @@ pub async fn download_release(
     let download_url =
         format!("https://github.com/moonrepo/proto/releases/download/v{version}/{download_file}");
 
+    trace!(
+        version,
+        triple,
+        "Downloading proto release from {}",
+        color::url(&download_url)
+    );
+
     // Request file from url
     let handle_error = |error: reqwest::Error| ProtoInstallerError::DownloadFailed {
         url: download_url.clone(),
@@ -68,6 +76,15 @@ pub async fn download_release(
         .send()
         .await
         .map_err(handle_error)?;
+
+    if !response.status().is_success() {
+        return Err(ProtoInstallerError::DownloadNotAvailable {
+            version: version.to_owned(),
+            status: Box::new(response.status()),
+        }
+        .into());
+    }
+
     let total_size = response.content_length().unwrap_or(0);
 
     on_chunk(0, total_size);
@@ -118,11 +135,23 @@ pub fn unpack_release(
         vec!["proto", "proto-shim"]
     };
 
+    trace!(
+        source = ?download.archive_file,
+        target = ?temp_dir,
+        "Unpacking downloaded proto release"
+    );
+
     // Unpack the downloaded file
     Archiver::new(&temp_dir, &download.archive_file).unpack_from_ext()?;
 
     // Move the old binaries
     let relocate = |current_path: &Path, relocate_path: &Path| -> miette::Result<()> {
+        trace!(
+            source = ?current_path,
+            target = ?relocate_path,
+            "Relocating old proto binary to a new versioned location",
+        );
+
         fs::rename(current_path, relocate_path)?;
 
         // Track last used so operations like clean continue to work
@@ -163,6 +192,8 @@ pub fn unpack_release(
 
     // Move the new binary to the bins directory
     let mut unpacked = false;
+
+    trace!(bin_dir = ?install_dir, "Moving unpacked proto binaries to the bin directory");
 
     for bin_name in &bin_names {
         let output_path = install_dir.join(bin_name);